requirements.txt
absl-py==2.4.0
annotated-types==0.7.0
ansible==13.6.0
ansible-core==2.20.5
anyio==4.13.0
asgiref==3.11.1
astroid==4.0.4
asttokens==3.0.1
astunparse==1.6.3
attrs==26.1.0
autopep8==1.5.5
backports.tarfile==1.2.0
bazel_runfiles==1.9.0
bleach==6.3.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
crcmod==1.7
cryptography==46.0.6
dbus-python==1.3.2
decorator==5.2.1
dill==0.4.1
distlib==0.4.0
distro==1.9.0
Django==5.2.13
dnspython==2.8.0
docstring_parser==0.17.0
executing==2.2.1
filelock==3.25.2
flake8==3.8.4
Flask==3.1.3
Flask-Login==0.6.3
Flask-PyMongo==3.0.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.12.19
gast==0.7.0
gce-v2v @ https://storage.googleapis.com/velostrata-release/gce-v2v/gce-v2v.tar.gz#sha256=1533c75fd582af8102f3c7bb800342521b0964266f723e29ebd5afddc1089c4d
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-cloud-aiplatform==1.143.0
google-cloud-appengine-logging==1.8.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-cloud-datastore==2.23.0
google-cloud-language==2.20.0
google-cloud-logging==3.15.0
google-cloud-monitoring==2.30.0
google-cloud-resource-manager==1.17.0
google-cloud-spanner==3.63.0
google-cloud-storage==3.10.1
google-cloud-translate==3.25.0
google-cloud-videointelligence==2.19.0
google-cloud-vision==3.13.0
google-crc32c==1.8.0
google-genai==1.68.0
google-pasta==0.2.0
google-resumable-media==2.8.2
googleapis-common-protos==1.73.1
greenlet==3.4.0
grpc-google-iam-v1==0.14.4
grpc-interceptor==0.15.4
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
isort==8.0.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
jedi==0.17.2
jeepney==0.9.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.13.2
keyring==25.7.0
keyrings.google-artifactregistry-auth==1.1.2
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libclang==18.1.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mccabe==0.6.1
mdurl==0.1.2
mercurial==6.7.2
ml_dtypes==0.5.4
mmh3==5.2.1
more-itertools==10.8.0
namex==0.1.0
numpy==2.4.4
oauth2client==4.1.3
oauthlib==3.2.2
opentelemetry-api==1.40.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
parso==0.7.1
pexpect==4.9.0
platformdirs==4.9.4
pluggy==1.6.0
prompt_toolkit==3.0.52
proto-plus==1.27.2
protobuf==6.33.6
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycodestyle==2.6.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydocstyle==6.3.0
pyflakes==2.2.0
Pygments==2.20.0
PyGObject==3.48.2
PyJWT==2.7.0
pylint==4.0.5
pymongo==4.16.0
pyparsing==3.3.2
python-apt==2.7.7+ubuntu5.2
python-dateutil==2.9.0.post0
python-discovery==1.2.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytoolconfig==1.3.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.0
resolvelib==1.2.1
rich==14.3.3
rope==1.14.0
rpds-py==0.30.0
rsa==4.9.1
SecretStorage==3.5.0
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
SQLAlchemy==2.0.49
sqlparse==0.5.5
stack-data==0.6.3
supervisor==4.2.5
tenacity==9.1.4
tensorflow==2.21.0
termcolor==3.3.0
tink==1.14.1
toml==0.10.2
tomlkit==0.14.0
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.6.3
virtualenv==21.2.0
wadllib==1.3.6
wcwidth==0.6.0
webencodings==0.5.1
websockets==16.0
Werkzeug==3.1.7
wheel==0.46.3
wrapt==2.1.2
yapf==0.43.0
zipp==3.23.0
Flask
Flask-SQLAlchemy
Flask-Security-Too